pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "google-auth>=2.40.3",
    "gspread>=6.2.1",
    "gunicorn>=23.0.0",
    "openai>=1.101.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.43",
]
